==> AWS/1. Data Cleaning (Basic).ipynb <==
### Data Exploration and Cleanup (basic) script
Our aim here is to understand our dataset and do a basic cleanup removing NaNs & Duplicates. 


!pip install boto3
import boto3
import pandas as pd
import io
import pandas as pd
import numpy as np
# Replace with your bucket and key
bucket_name = 'demand-prediction-ola-rides'
file_key = 'raw_data.csv'  # e.g. 'data/mydata.csv'

# Create a Boto3 client
s3 = boto3.client('s3')

# Read the file content into a BytesIO object
response = s3.get_object(Bucket=bucket_name, Key=file_key)
content = response['Body'].read()




# Convert to DataFrame
df = pd.read_csv(io.BytesIO(content), compression='gzip')

# Display the first few rows
df.head(5)

!ls

#### Reading DataSet

len(df)

### A Customer_ID `number` at a particular timestamp can only have one entry
### Removing Duplicate Entries ['ts','number']

df[df.duplicated(subset=['ts','number'],keep=False)]

### There are 113540 Duplicate Entries
#### We have 8315498 Unique timestamp, customer_id rows. 

## Keeping first occurence
df.drop_duplicates(subset=['ts','number'], inplace = True, keep = 'last')

df.reset_index(inplace = True, drop = True)

# Info of Dataset
df.info()

#Count missing values
np.count_nonzero(df.isnull().values)

df['number'] = pd.to_numeric(df['number'], errors = 'coerce')

#Count missing values
np.count_nonzero(df.isnull().values)

#### There are 116 NaN rows, dropping NaN rows.

df.dropna(inplace = True)
len(df)

df['number'] = pd.to_numeric(df['number'], errors = 'coerce', downcast='integer')
df['ts'] = pd.to_datetime(df['ts'])

# Info of Dataset
df.info()

### Breaking Time to Features

df['hour'] = df['ts'].dt.hour
df['mins'] = df['ts'].dt.minute
df['day'] = df['ts'].dt.day
df['month'] = df['ts'].dt.month
df['year'] = df['ts'].dt.year
df['dayofweek'] = df['ts'].dt.dayofweek

import io
import boto3

# Assuming df is your DataFrame and s3 client and bucket_name defined

buffer = io.BytesIO()
df.to_csv(buffer, index=False, compression='gzip')  # <-- Add compression here
buffer.seek(0)  # Reset pointer to start

# Upload to S3
s3.put_object(Bucket=bucket_name, Key='processed/preprocessed_11.csv.gz', Body=buffer.getvalue())

print("In-memory gzip-compressed CSV uploaded successfully!")


print(f"Size in MB: {response['ContentLength'] / (1024*1024):.2f}")


df.head(5)
