==> AWS/5. Prediction Pipeline.ipynb <==
### Task: Given 26th March 2021 demand - clean dataset based on business team rules; we need to forecast demand for first 3 (30mins internal) demand value of 27th March 2021.

Author: [Shaurya Uppal](https://www.linkedin.com/in/shaurya-uppal/)

import pandas as pd
import numpy as np
from joblib import load, dump
from sklearn.cluster import MiniBatchKMeans, KMeans
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error
from math import sqrt, ceil, floor
from datetime import datetime, timedelta

def round_timestamp_30interval(x):
    if type(x)==str:
        x = datetime.strptime(x, '%Y-%m-%d %H:%M:%S')
    return x- timedelta(minutes=x.minute%30, seconds=x.second, microseconds=x.microsecond)

def time_features(data):
    data['mins'] = data.ts.dt.minute
    data['hour'] = data.ts.dt.hour
    data['day'] = data.ts.dt.day
    data['month'] = data.ts.dt.month
    data['dayofweek'] = data.ts.dt.dayofweek
    data['quarter'] = data.ts.dt.quarter
    return data

def prediction_without_lag(df):
    return predict_without_lag.predict(df[['pickup_cluster','mins','hour','month','quarter','dayofweek']])

def prediction_with_lag(df):
    return predict_with_lag.predict(df[['pickup_cluster', 'mins', 'hour', 'month', 'quarter',
           'dayofweek', 'lag_1', 'lag_2', 'lag_3','rolling_mean']])

def shift_with_lag_and_rollingmean(df):
    df = df.sort_values(by=['pickup_cluster', 'ts']).drop_duplicates(subset=['ts','pickup_cluster'])
    df = df.set_index(['ts', 'pickup_cluster', 'mins', 'hour', 'month', 'quarter', 'dayofweek'])
    df['lag_1'] = df.groupby(level=['pickup_cluster'])['request_count'].shift(1)
    df['lag_2'] = df.groupby(level=['pickup_cluster'])['request_count'].shift(2)
    df['lag_3'] = df.groupby(level=['pickup_cluster'])['request_count'].shift(3)
    df['rolling_mean'] = df.groupby(level=['pickup_cluster'])['request_count'].apply(lambda x: x.rolling(window = 3).mean()).shift(1)

    df = df.reset_index(drop = False).dropna()
    df = df[['ts', 'pickup_cluster', 'mins', 'hour', 'month', 'quarter',
           'dayofweek', 'lag_1', 'lag_2', 'lag_3','rolling_mean','request_count']]
    return df



df = pd.read_csv('../Data/test_dataset/cleaned_test_booking_data.csv', compression = 'gzip', low_memory=False)
cluster_model = load('../Model/pickup_cluster_model.joblib')
predict_without_lag = load('../Model/prediction_model_without_lag.joblib')
predict_with_lag = load('../Model/prediction_model.joblib')


### Use Clustering Kmeans Model for Geospacial Feature - `pickup_cluster`

df['pickup_cluster'] = cluster_model.predict(df[['pick_lat','pick_lng']])
df.head(10)

### Data preparation and processing

df['ts'] = np.vectorize(round_timestamp_30interval)(df['ts'])
df['ts'] = pd.to_datetime(df['ts'])

df = df[['ts','number','pickup_cluster']]
df=df.groupby(by = ['ts','pickup_cluster']).count().reset_index()
df.columns = ['ts','pickup_cluster','request_count']

## Adding Dummy pickup cluster -1

## Change this Data based on your data
l = [datetime(2021,3,26,00,00,00) + timedelta(minutes = 30*i) for i in range(0,51)]
lt = []
for x in l:
    lt.append([x, -1, 0])
temp = pd.DataFrame(lt, columns = ['ts','pickup_cluster','request_count'])
df = df.append(temp,ignore_index=True)

data = df.set_index(['ts', 'pickup_cluster']).unstack().fillna(value=0).asfreq(freq='30Min').stack().sort_index(level=1).reset_index()

# Removing Dummy Cluster
data = data[data.pickup_cluster>=0]

df = time_features(data)

### Model without Lag (past data) requirement

data_without_lag = df[df['ts']>=datetime(2021,3,27,00,00,00)].__copy__()
data_without_lag['request_count'] = prediction_without_lag(data_without_lag)
data_without_lag

data_without_lag.to_csv('../Data/test_dataset_prediction_output/prediction_without_lag_model.csv',index = False, compression = 'gzip')

### Using Iteration 3 - Best Model with Lag Features and Rolling Means (Recursive Multi-Step Forecast used)

start_date = datetime(2021,3,27,00,00,00) 
for x in range(3):
    df = shift_with_lag_and_rollingmean(df)
    df.loc[df[df['ts']==start_date+timedelta(minutes=30*x)].index,'request_count'] = prediction_with_lag(df[df['ts']==start_date+timedelta(minutes=30*x)])

data_with_lag = df[df['ts']>=datetime(2021,3,27,00,00,00)].__copy__()
data_with_lag

data_with_lag.to_csv('../Data/test_dataset_prediction_output/prediction_with_lag_model.csv',index = False, compression = 'gzip')

### Thank You for Choosing this Project. Hope you liked it. Check our Scripts Folder for Modularized py scripts.
